requirements.txt
streamlit
pandas
mysql-connector-python
sqlglot
scikit-learn
joblib
